requirements.txt
fastapi
pydantic
loguru
uvicorn
EdgeGPT
pyyaml